Update GenBankTabler.py
Now includes function to tell it what to do with voucherless specimens, which bloat the tables. --include_unlinked outputs voucherless_*.csv for each marker that has voucherless specimens. If not specified, these individuals are not included in the output.

Known issue: does not yet check to merge identical individuals where Isolate and Voucher are the same information.