Update GenBankTabler.py
Added function to set a limit for 10000 sequences so that you do not accidentally download all of GenBank. 
Also added a flag to warn you how many sequences you are downloading and gives you the option to cancel. Plus, adds a flag to auto-confirm so that batch runs can be processed.