Update GenBankTabler.py
Added standardised names for common gene aliases, and fuzzy matching to gene names, to account for inconsistencies in formatting and uploading. These can be overriden by a user input .csv file with the new --marker_aliases flag containing canonical names in the first column and aliases in the second column, one per line.